Changed the signature of \Break
Instead of taking atom occurrence labels it now takes atom occurrences and the order of the arguments has changed.
This might be a bit more complicated to read, but I think it is counterintuitive to use labels directly. My understanding is that unless we explicitly put labels on atom occurrences (for our convenience), atom occurrences do not have labels. So referring to atom occurrences by their labels does not make sense.