Defined what it means to be a path from an edge to an edge and added TODO
TODO: Add the definition of 'upper' and 'lower' edges of vertices and flows. say what it means to be an edge from x to y. at the moment a path from x to y is not a path from y to x unlike in AFI. this we can fix, but we should check the rest of the paper.